+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated optional mods on cursegradle
 - Refactored some classes
 - Skip plugin load before try catch
+- Updated Applied Energistics Plugin
+    - Add in world fluid and seed crafting alongside condenser singularity and matter_ball
 
 ### Removed
 - Harvestcraft jar, we fetch from jitpack now